audio: 16-bit unipolar SN76489 encoding with headroom (#82 Step D/E)
The 8-bit unipolar samples (0..254) had no headroom, so the anti-alias
filter's ~11% step overshoot clipped full-volume tones (about 6% of samples
pinned at the rail) and the 8-bit floor cost ~48-60 dB of SNR versus 16-bit.
Move the SN76489 to two ENCODING_2X16BIT_SIGNED source fields (source 0 =
tone0,tone1; source 1 = tone2,noise), unipolar, silence 0, full-scale high
level Sn76489::FULL_SCALE = 16384 -- one named constant, quoted in the proto
and docs. Levels come from the -2 dB/step law in floating point (not scaled
up from the 7-bit table), decimated in double and clamped to int16 as a guard
that no single channel engages.

All in-repo producers and consumers move together (no client/server back-compat):
- core: emit 2x16; GetAudioFormat describes two SN sources with MOS channel
  names and one group; ENCODING_4X8BIT_UNSIGNED kept but marked unused.
- macOS: ring buffer carries a 64-bit frame (both source words); AudioClient
  extracts sources 0+1; AudioRenderer unpacks four int16 and normalises by half
  full-scale so a volume-0 square keeps unit AC amplitude (loudness unchanged),
  then high-passes the DC. Builds clean.
- Python: audio.py comment; a client test unpacks a chunk per GetAudioFormat.
- proto comments regenerate the Python stubs (.pyi); no schema change.
- docs/sound-subsystem.md and the analysis tools updated to 2x16.
- C++ sound tests unpack via a shared tests/sn76489_channels.hpp helper and
  assert on the int16 full-scale model.

The clipping guard now passes (peak rides above full scale, no rail hits);
baseband recovery and out-of-band remain green.

+++ clients/macos/Beebium/Beebium/AudioRingBuffer.swift
@@ -1,4 +1,4 @@
-// Copyright 2025 Robert Smallshire <robert@smallshire.org.uk>
+// Copyright 2026 Robert Smallshire <robert@smallshire.org.uk>
 //
 // This file is part of Beebium.
 //
@@ -15,8 +15,9 @@ import Atomics
 
 /// Thread-safe lock-free single-producer single-consumer (SPSC) ring buffer for audio samples.
 ///
-/// The buffer stores packed 32-bit audio samples where each sample contains 4 x 8-bit signed
-/// channel values (Tone0, Tone1, Tone2, Noise) in big-endian order within the 32-bit field.
+/// The buffer stores one 64-bit element per audio frame: the SN76489's two
+/// 32-bit source fields (source 0 = tone0|tone1, source 1 = tone2|noise), each
+/// two signed 16-bit channels, packed with source 0 in the low half.
 ///
 /// Thread safety:
 /// - Producer thread (gRPC callback): calls `write()` to add samples
@@ -29,7 +30,10 @@ final class AudioRingBuffer: @unchecked Sendable {
     /// Default capacity in samples (~85ms at 48kHz)
     static let defaultCapacity = 4096
 
-    private let buffer: UnsafeMutableBufferPointer<UInt32>
+    // Each element holds one audio frame: the SN76489's two 32-bit source
+    // fields (source 0 = tone0|tone1, source 1 = tone2|noise) packed as a
+    // UInt64, low half = source 0.
+    private let buffer: UnsafeMutableBufferPointer<UInt64>
     private let capacity: Int
 
     /// Read index - only modified by consumer, read by producer
@@ -67,7 +71,7 @@ final class AudioRingBuffer: @unchecked Sendable {
     ///   - samples: Pointer to packed 32-bit samples
     ///   - count: Number of samples to write
     /// - Returns: Number of samples actually written (may be less if buffer full)
-    func write(_ samples: UnsafePointer<UInt32>, count: Int) -> Int {
+    func write(_ samples: UnsafePointer<UInt64>, count: Int) -> Int {
         let currentWrite = writeIndex.load(ordering: .relaxed)
         let currentRead = readIndex.load(ordering: .acquiring)
 
@@ -107,15 +111,15 @@ final class AudioRingBuffer: @unchecked Sendable {
         return toWrite
     }
 
-    /// Write samples from a Data object containing packed UInt32 values.
+    /// Write samples from a Data object containing packed UInt64 frames.
     ///
-    /// - Parameter data: Data containing packed samples (4 bytes per sample)
+    /// - Parameter data: Data containing packed frames (8 bytes per frame)
     /// - Returns: Number of samples written
     func write(data: Data) -> Int {
-        let sampleCount = data.count / MemoryLayout<UInt32>.size
+        let sampleCount = data.count / MemoryLayout<UInt64>.size
         return data.withUnsafeBytes { rawBuffer in
             guard let baseAddress = rawBuffer.baseAddress else { return 0 }
-            let samples = baseAddress.assumingMemoryBound(to: UInt32.self)
+            let samples = baseAddress.assumingMemoryBound(to: UInt64.self)
             return write(samples, count: sampleCount)
         }
     }
@@ -131,7 +135,7 @@ final class AudioRingBuffer: @unchecked Sendable {
     ///   - destination: Buffer to write samples into
     ///   - count: Maximum number of samples to read
     /// - Returns: Number of samples actually read
-    func read(_ destination: UnsafeMutablePointer<UInt32>, count: Int) -> Int {
+    func read(_ destination: UnsafeMutablePointer<UInt64>, count: Int) -> Int {
         let currentRead = readIndex.load(ordering: .relaxed)
         let currentWrite = writeIndex.load(ordering: .acquiring)
